requirements.txt
gomatic==0.3.2
PyGithub==1.25.2
